chore: reducing the amount of instance calls and of course a little bit of cleaning

+++ src/main/java/dev/revere/alley/api/menu/Menu.java
@@ -102,18 +102,18 @@ public abstract class Menu {
      * @param data    the data
      * @param rows    the rows
      */
-    public void addBorder(Map<Integer, Button> buttons, byte data, int rows) {
+    public void addBorder(Map<Integer, Button> buttons, int data, int rows) {
         for (int i = 0; i < 9; i++) {
-            buttons.putIfAbsent(i, Button.placeholder(Material.STAINED_GLASS_PANE, data, ""));
+            buttons.putIfAbsent(i, Button.placeholder(Material.STAINED_GLASS_PANE, (byte) data, ""));
         }
 
         for (int i = 0; i < 9; i++) {
-            buttons.putIfAbsent((rows - 1) * 9 + i, Button.placeholder(Material.STAINED_GLASS_PANE, data, ""));
+            buttons.putIfAbsent((rows - 1) * 9 + i, Button.placeholder(Material.STAINED_GLASS_PANE, (byte) data, ""));
         }
 
         for (int i = 0; i < rows; i++) {
-            buttons.putIfAbsent(i * 9, Button.placeholder(Material.STAINED_GLASS_PANE, data, ""));
-            buttons.putIfAbsent(i * 9 + 8, Button.placeholder(Material.STAINED_GLASS_PANE, data, ""));
+            buttons.putIfAbsent(i * 9, Button.placeholder(Material.STAINED_GLASS_PANE, (byte) data, ""));
+            buttons.putIfAbsent(i * 9 + 8, Button.placeholder(Material.STAINED_GLASS_PANE, (byte) data, ""));
         }
     }
