+++ src/main/java/dev/revere/alley/api/menu/pagination/impl/menu/ViewAllPagesMenu.java
@@ -37,7 +37,7 @@ public class ViewAllPagesMenu extends Menu {
             }
         }
 
-        this.addBorder(buttons, (byte) 15, 5);
+        this.addBorder(buttons, 15, 5);
 
         return buttons;
     }
